+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 <!-- There should always be "Unreleased" section at the beginning. -->
 
 ## Unreleased
+- Profile used endpoint details for a query
 
 ## 1.0.0 - 2021-05-13
 - Initial implementation